MINOR rename package from migration to wrapper
Change to package name wrapper, since we plan to have a few
other serdes that wrap others
